Android tweaks

+++ code/mob/invProcs.dm
@@ -5,13 +5,19 @@ mob
 		}
 
 		addInv(obj/item/o){
-			if(!o.dBID){o.dBID = (getLastID() + 1);}
-			_query("INSERT INTO `inventory` (`ID`,`type`,`owner`) VALUES ('[o.dBID]','[o.type]','[name]');");
+			if (istype(src, /mob/Player)) {
+				if(!o.dBID){o.dBID = (getLastID() + 1);}
+				_query("INSERT INTO `inventory` (`ID`,`type`,`owner`) VALUES ('[o.dBID]','[o.type]','[name]');");
+			}
 			o.Move(src)
 			qFac.checkItem(src)
 		}
 
 		remInv(obj/item/o){
+			// Don't SQL for mobs
+			if (istype(src, /mob/NPA)) {
+				return
+			}
 			_query("DELETE FROM `inventory` WHERE `ID`='[o.dBID]';");
 		}
 
@@ -27,9 +33,15 @@ mob
 			if(o.DECAY) o.decayItem(o.loc)
 		}
 
-		giveItem(obj/item/o, mob/Player/m){
-			remInv(o)
-			m.addInv(o)
+		giveItem(obj/item/O, mob/m){
+			send("You give [m.raceColor(m.name)] [O:PREFIX][O:DISPLAY].",src)
+			send("[src.raceColor(src.name)] gives you [O:PREFIX][O:DISPLAY].", m)
+			send("[src.raceColor(src.name)] gives [m.raceColor(m.name)] [O:PREFIX][O:DISPLAY].",a_oview_extra(0,src,m))
+			remInv(O)
+			m.addInv(O)
+			if (istype(m, /mob/NPA)) {
+				m.receive_item(O, src)
+			}
 		}
 
 		equipItem(obj/item/o,mob/Player/m){
